added optionals and first working test
